[audio_core] Update clock terminology
AudioCore is updated to consistently use just the "adjustable"
and "non-adjustable" labels for AudioClocks (in addition to
their source: "client" and "device"). Other terms are removed,
except the term "flexible" which is still used sparingly in
the directly-client-facing renderer and capturer classes.

In other CL, "flexible" replaces "optimal" or other
terms in documentation, examples and public headers when
describing the clock supplied (and adjusted) by AudioCore.
Although internally we refer to such a clock as "adjustable"
(from AudioCore's point-of-view), that would be entirely
confusing to clients since it is a clock that they CANNOT adjust.

Test: No behavior change, naming changes only

Change-Id: I4893091763767125b4c09acabcfa0b1da41b1097
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434102
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Testability-Review: Tom Bergan <tombergan@google.com>