[brcmfmac] C++ style fixes for cfg80211.
This change fixes some of the clang_tidy warnings for this file, but not
all of them because the resulting CL would be too large. Included fixes
are: modernize-loop-convert and readability-else-after-return.

Test: this is not a functional change. fx test brcmfmac-test and the
ACTS test suite WlanPhyComplianceABGTest continue to pass with this
change.

Change-Id: I1d778e20717c9b61e1b5c5c9bb6a29ae8dc7211b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434279
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Testability-Review: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Karl Ward <karlward@google.com>