[lsan] Fixes LSAN memory leak in display controller
Fixes the lifetime of a watcher closure, which under some circumstances
could cause a reference to the closure not to be deallocated.  The old
code employs a clever trick to ensure that a reference to a closure
expires when the closure is called.  However, it does not release memory
if the closure is never called, leading to a memory leak.

The new code ensures that the reference is removed when either this goes
out of scope, or the closure is called.  The approach is not thread
safe, and is assuming that the async loop is attached to the main
thread.

Fixed: 61246
Change-Id: I0a72131048f01e1f0925bd15c584b96c333bfe3d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434517
Reviewed-by: Josh Gargus <jjosh@google.com>
Testability-Review: Filip Filmar <fmil@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>