[kernel][acpi_lite] add new lightweight library to parse necessary kernel ACPI tables
Add a simple library that replaces the simple part of finding the root
tables and decoding the simple ones that are used in the kernel.

Next change will wire it up to the acpi tables library which has test
code validating that the low level interface is working.

Testing: validate that it finds the ACPI table on a variety of hardware
including qemu and a few AMD and Intel machines.

Bug: 32149
Change-Id: Ic798bb950d49904987953a010de5ecba5ffa2819
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427694
Commit-Queue: David Greenaway <dgreenaway@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>
Testability-Review: Adrian Danis <adanis@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>