[devices] Request private clones of driver binary
This is part of the drivers-as-components work. As of fxrev.dev/429697,
memfs will only set ZX_RIGHT_SET_PROPERTY on private clones of VMOs.
When requesting the VMO of a driver binary, request a private clone so
that we can set the name of the VMO for accounting purposes. As bootfs
is backed by memfs, without requesting a private clone we can not name a
VMO.

Bug: 33183
Change-Id: Ie49234555e769d561a04843db7b9aace012be058
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435066
Commit-Queue: Abdulla Kamar <abdulla@google.com>
Testability-Review: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>