[sys/core] put storage in "persistent" subdir
Modify the storage capability declaration for "data" to use a subdir
with the name "persistent".

This is a breaking change, and users of this storage capability will
lose access to their old storage due to this change. Archivist is the
only such component, and can handle this.

Change-Id: If0f9f5109b27b7f4d0c9bac81175be2a1e30f2a9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434194
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Testability-Review: Gary Bressler <geb@google.com>
Commit-Queue: Derek Gonyeo <dgonyeo@google.com>