[wlan][sme] Log previous connected AP to Inspect
If a bug report is uploaded while a user is not connected wirelessly, or
while a user is connected to a different AP, we would lose information
for debugging the previous disconnect. In this commit, we log to Inspect
the previously connected AP once SME moves to the idle state from a
connected state.

Bug: 60154
Test:
1. Manual verification
2. fx test wlanstack-tests
3. fx test fuchsia-pkg://fuchsia.com/wlan-hw-sim-test#meta/verify-wlan-inspect.cmx

Change-Id: If2638fac2ef8bc2471cbf8ee205716ce81d6ec40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430314
Commit-Queue: Kiet Tran <kiettran@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>
Testability-Review: Charles Celerier <chcl@google.com>