[flatland] Default FlatlandPresenter impl.
The FlatlandPresenter interface was primarily added to:
1. Isolate Flatland from having a direct dependency on the
   FrameScheduler.
2. Introduce a natural location to add a thread-safety later around
   the FrameScheduler (rather than enforcing thread-safety within the
   FrameScheduler itself).

As a result, the simplest implementation of the DefaultFrameScheduler
simply places FrameScheduler calls into a task posted onto the main
looper. A lock would be simpler, but GFX clients do not access the
FrameScheduler through the same object, meaning a lock is not
sufficient when Flatland and GFX are running simultaneously.

Future changes in the FrameScheduler could eliminate the need for the
additional thread-safety logic.

TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Change-Id: I017c386157c05c668e0dee6641140ad9c2061434
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434038
Commit-Queue: Evan Birenbaum <birenbaum@google.com>
Reviewed-by: Adam Gousetis <adamgousetis@google.com>
Reviewed-by: Mikael Pessa <mikaelpessa@google.com>
Testability-Review: Mikael Pessa <mikaelpessa@google.com>