[ffx] Drop futures when FIDL client stream drops.
This fixes up a bit of a hack from a previous CL so that there aren't
any wrapped errors. This is identical to `try_for_each_concurrent`
except that futures are dropped when the client disconnects and the end
of the stream is reached.

This does so with `try_for_each_concurrent_while_connected` which
handles requests so long as the end of the client stream has not been
reached.

Change-Id: Ief641f7590c6004a45fe709de46d1cd9d4992ce7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435055
Commit-Queue: Andrew Davies <awdavies@google.com>
Testability-Review: Matthew Boetger <boetger@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>