[core][bringup] Support appmgr resolution failure
Disable the package resolver in the bringup product since bringup
is not expected to support blobfs and pkgfs.

Introduce the 'core starter' component which is in charge of
starting appmgr. Without the package resolver, resolution of
appmgr fails. With the previous topology this was fatal to the
system because appmgr was in a direct `eager` lineage from the
root component and its failure to bind cascaded to the root.
The core starter becomes the direct, `eager` descendant of root,
attempts to use a capability from appmgr, and exits, regardless
of result.

Fixed: 58182

TEST=startup-integration-test

Change-Id: I0a76ceda4b3b857d25d9e85e9b861ab4330a963e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428040
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Commit-Queue: Justin Mattson <jmatt@google.com>
API-Review: Gary Bressler <geb@google.com>
Testability-Review: Gary Bressler <geb@google.com>