[i915] Trivial: Simplify `Controller::AddDisplay`.
Currently, we reserve a single slot in the fbl::Vector each time
`AddDisplay` is called. Each such `Vector::reserve` operation will need
to perform an allocation and move all existing elements to this new
allocation. The result is that adding `n` displays currently requires
O(n**2) operations.

While the number of displays is likely to be small (so n**2 is also
likely to be small), it is also simpler to just not reserve the space.

Simplify the implementation a little.

Change-Id: I2b2042a71e694e6fab5d2a5447cd6c186c824403
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435063
Testability-Review: David Greenaway <dgreenaway@google.com>
Reviewed-by: Randall Bosetti <rlb@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>