[kernel][arm][gicv3] Read the correct max SPI field
Previously was reading a field that means something else, but was
probably always larger than necessary on all of the socs that have
been tested on previously.

Updated the GICv2 detection to print a similar message.

Tested: boot on a few pieces of hardware, validating that the interrupt
count that is read from the register lines up with reality.

Change-Id: I963e2edf59e273660f7924bd84fec20a24f8a399
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/424033
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Testability-Review: Adrian Danis <adanis@google.com>