[kernel][oom] Ignore dlog_shutdown result when OOMing
When rebooting because of an OOM condition, don't assert on the result
of dlog_shutdown.  Failing the assert results in a kernel panic and
misclassification of the reboot cause.  Two possible reasons for
dlog_shutdown to fail are 1) the dlog subsystem is already in the
process of being shutdown and 2) CPU starvation has prevented the
already queued dlog messages from being written to the serial port
before the 20 second timeout has elapsed.  In either of these cases
there's not much else we can do that's better than simply continuing
on with the reboot.

Fix and update adjacent code comments.

Fixed: 55673
Change-Id: Iad33c0b6dc7463d8f1f9f0c6b5631380bb801570
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435435
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>