[fx] Make `fx list-build-artifacts` building explicit
Now that the name has been changed to "list", the implicitly behavior of
building is awkward. This change replaces `--no-build` with `--build`.

This will require an update of a local checkout w.r.t to the
work-from-home workflow, but that is already required of the parent
renaming change, so it minimizes disruption to do this now.

Change-Id: Ib7f9f4ac00ad7eae5836e6b396319414d6d94244
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435575
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Testability-Review: Renato Mangini Dias <mangini@google.com>