[wlan][policy] hidden network config support
This change adds a field for the estimated probability that a saved
network is hidden. For now the value will always be 0, meaning we
shouldn't actively probe for the network. Another field is also added
to network configs for deciding whether the network is hidden: have
we connected to this network using a passive scan. It also adds
methods to the SavedNetworksManager to update configs.

Multiply: wlancfg-tests
Bug: 35916
Test: Added unit tests
	fx test wlancfg-tests
Change-Id: I41debe9d14b42758ab323aff2ecba8ead572c138
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/430795
Commit-Queue: Naomi Mccracken <nmccracken@google.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>
Testability-Review: Kevin Sakuma <sakuma@google.com>