[doc_checker]Allowing http link to OWNERS files
This file is in the //docs path, and the general rule is
http based links are not allowed to files under //docs, they should be
file paths starting at /docs.

The OWNERS file is a special case since it is not markdown, it is
not copied to fuchsia.dev for publishing, so it should be a link
back to the source code.

This fix adds a list of allowed paths to handle this situation.

Change-Id: I278ffe61fd5dac28d67b78fcd80808fe61865814
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436054
Testability-Review: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>