[build][fuzzing] Fix logging for no-Go fuzzers
fxr/412795 encountered trouble building go fuzzers with its changes to
the syslog backend. It tried to work around this by building the
libraries as static when in a fuzzer toolchain. This did build, but the
resulting fuzzer binaries had weak, unresolved symbols for the logging
macros, leading to crashes.

This CL isn't a fix for the Go fuzzer linking issue. Instead, it narrows
the workaround to only apply to go fuzzers, and to forbid them from
having a dep on //sdk/lib/syslog/cpp. This allows no-Go fuzzers to
depend on syslog in the expected manner.

Bugs: 44551, 59937, 60948, 61348
Change-Id: I1f5fad030bf2c91cfa800f1c71834bb558a11a56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436074
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>