[kernel][acpi_lite] Fix bug parsing XSDT tables.
Fix bug where only the first half of tables in the XSDT can be read, and
such tables can only be read if in the first 4GiB of memory.

The bug is because we incorrectly treat the list of physical addresses in
the XSDT as being an array of 32-bit values instead of 64-bit values.
This means that every 64-bit entry will be read as two 32-bit entries,
which will typically be a 0 entry (the high 32-bits) followed by a real
entry (the low 32-bits).

This was not caught on existing tests because ACPI tables the kernel
needs on systems using XSDT are early on in the ACPI tables.

Added new tests, and verified on a Chromebook Atlas all entries are now
being read.

Change-Id: Id98e1f066276137032a73e5832537f4fab744fcb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436494
Testability-Review: David Greenaway <dgreenaway@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>