[tools] Use correct target for triage and add tool_paths metadata.
The previous target was always installing the tool with the host
toolchain, not the linux x64 toolchain that we want. Also, the
tool_paths metadata was only getting added if the current toolchain=
host_toolchain, so although it would build the linux x64 version of the
tool, it wouldn't record the path in the tool_paths.json. This change
will add the metadata if the current toolchain is for any host, not just
the current host.

Bug: 57570
Change-Id: Ibad580d826507887760fd9fc4299a686ecddf9be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436195
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>