[gidl][hlcpp] Separate FIDL generation per GIDL file
Currently, the HLCPP benchmarks are built within the same
source set, with each benchmark importing a large file
with all of the FIDL headers. This follows what is done
for the rust benchmarks so a minimal number of FIDL files
is used for each GIDL file.

This does not have significant effects on build time, but
it does move towards a leaner build configuration and the
output filenames now match the specific GIDL test -- making
it easier to understand what is going on.

Change-Id: I3c2d0818bde4e663aa9d7cfc09b640f60962d134
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435843
Reviewed-by: Mitchell Kember <mkember@google.com>
Testability-Review: Mitchell Kember <mkember@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>