[blobfs] Implement the BlobLayout class
The BlobLayout class defines where the Merkle tree and data are
positioned within a blob for different formats.  Additional methods are
defined to consolidate the calculations performed on the layout to a
single testable location.  This avoids having to worry about underflows
and overflows when dealing with the layout in the rest of blobfs.

Bug: 36663
Change-Id: Ifaa666c6fce1fed7b5b0a6599e58520558b5dcd6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429216
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Testability-Review: Gianfranco Valentino <gevalentino@google.com>
Testability-Review: James Sullivan <jfsulliv@google.com>