[graphics][compute] Remove set_sources_assignment_filter() use from build files.
This GN feature is up for removal in future versions, so implement what
is needed with simple file-extension filtering instead.

This should not change the build. For verification, the content of
out/default has been compared before/after the CL is applied to verify
it is identical.

Bug: 61489
Testing: Should not change the build. See above.
Change-Id: I680f67d8292d020ca7ac57cf702e2f9791cc2ecf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436340
Reviewed-by: Allan MacKinnon <allanmac@google.com>
Commit-Queue: David Turner <digit@google.com>