[time] Report the median time from HTTPSDate
Adds a half second to the time reported by an HTTP server. Assuming
that the server truncates the actual UTC time to produce the value in
the date header, the half second is the median of the possible
subsecond values.

Change-Id: Ie79a5421602f2f44e16caebe4028406ea30ed656
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435741
Reviewed-by: Jody Sankey <jsankey@google.com>
Testability-Review: Jody Sankey <jsankey@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>