[brcmfmac] Add BRCMF_DBG_EVENT macro to make brcmf_event_msg logging consistent
There are a number of handlers used to process various events sent to
the driver from the firmware. This change makes event handler log the
contents of the event that triggered it to be logged in a consistent
manner. Note that logging each event from within each handler is
necessary because the reason code in each event message can depend on
the type of event.

Test: None necessary. Logging change only.
Change-Id: Iad0e8b4f30ec3a4203d51b6b7a4d2eb366c18794
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434916
Commit-Queue: Charles Celerier <chcl@google.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Testability-Review: Zhiyi Chen <zhiyichen@google.com>