[flatland] Wire up buffer collection release fences.
When Flatland::DeregisterBufferCollection() was originally
implemented, deregistration with the Renderer was hooked up to a
release fence stored in the Flatland object in anticipation of
integration with the FrameScheduler. Now that Flatland accepts
release fences from clients and forwards them to the
FlatlandPresenter, having separate logic for this fence isn't
appropriate.

This change modifies Flatland to only make a special release fence
when there are actually buffers ready for release in a particular
Present(). This fence is appended to the to the list of client
provided release fences so it is signaled at the same time: when the
Present() has been fully processed.

Fixed: 53330
TEST=fx test -o flatland_unittests
MULTIPLY: flatland_unittests (fuchsia): 30

Change-Id: I9f11bb0d2b08c2bb77d0c84e1ceb2aa548cd11f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435716
Commit-Queue: Evan Birenbaum <birenbaum@google.com>
Reviewed-by: Mikael Pessa <mikaelpessa@google.com>
Testability-Review: Mikael Pessa <mikaelpessa@google.com>