[gidl][llcpp] Separate FIDL generation per GIDL file
Currently, the LLCPP benchmarks are built within the same
source set, with each benchmark importing a large file
with all of the FIDL headers. This follows what is done
for the rust benchmarks so a minimal number of FIDL files
is used for each GIDL file.

This does not have significant effects on build time, but
it does move towards a leaner build configuration and the
output filenames now match the specific GIDL test -- making
it easier to understand what is going on.

This is the LLCPP equivalent of a recent HLCPP change.

Change-Id: Id992ed3814f59c68f9816c32f1f6f4d5679359ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436554
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Testability-Review: Mitchell Kember <mkember@google.com>