[timekeeper] Prepare to add HTTPSdate as a monitor source.
After this CL it should be a one line change to actually launch a
monitor, but the first time I actually tried launching httpsdate
was crashing.

This adds some basic logging for the monitor but will be much
more useful when per-time-source inpect is added in the next CL.

Note this also includes a couple of trait implementations (e.g.
Debug for a time source) that are needed in the next CL.

Bug: 56858
Change-Id: I2364310a7d8a00f25fc32b291b6ce843d57d763d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435840
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>
Testability-Review: Satsuki Ueno <satsukiu@google.com>