[storage][extractor] Add extractor library 3/n
The library in its current state provides a way to write an image
file. Storage software can use this library to dump the data that is
relevant to them.

The library is in rust. It provides C bindings(in c-bindings directory)
that can be used in C++ code base.

This patch adds image writer part of the library.

Test: Ran host side unit tests

Change-Id: I65e48700b482ad5ba065236101e849aa93f55ce6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432635
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Vikram Auradkar <auradkar@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>