[fdf] Add a binding header target for vendors
This is part of a transition to an allowlist to prevent new uses of the
old bind rules. For other repositories (i.e. vendor repositories) we
need an escape hatch since we can't add them to the allowlist. This must
be set up first so that those repositories can transition before we set
up the allowlist.

Change-Id: Ibf72203267b09ea8a23d72b07facbc8ee4cae46a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436757
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Alex Legg <alexlegg@google.com>