[bt][rfcomm] Update Session to use closure for opened channels
Updates Session to used a channel opened callback, instead of directly
passing in the set of Clients. This clarifies the abstraction between an
RFCOMM Session, and the clients of the RfcommServer. Now, Session does
not need to know anything about its clients, and can simply deliver
opened RFCOMM channels to the provided closure.

This simplifies some of the tests in session.rs, and removes the
dependency on FIDL.

Bug: 59582
Test: Updated bt-rfcomm-tests
Change-Id: I98114c7087fd80706aba55673a7e648c5b92e4c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434814
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Testability-Review: Lucas Jenkins <lucasjenkins@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>