[intl] Factors intl configuration out of basemgr
Removes the responsibility of serving fuchsia.intl.PropertyProvider
from basemgr.

This allows us to provide locale information for lower-level services
and brings the configuration of "low disk" devices closer to the
default.

Replaces the condition under which basemgr is started up - it seems
that on systems that take in basemgr.config there isn't a way to
start up basemgr automatically, so we do it this way.

Some background:

This is a 2nd attempt to merge, see revert at: http://fxr/436339.
The root cause for the revert was the change to `modular.config`, which
removed the only remaining stanza that forced `basemgr.cmx` to start
at system startup: there used to be more, but those apparently were
removed, accidentally leaving this one to be last.  In this edition
the change to `modular.config` was modified to use a different service
for eager startup of `basemgr.cmx`.

This change was then verified manually on emu, since no current
automated tests verify that basemgr is up and running.  In a separate
change I'm working on an E2E test that will guard basemgr startup
behavior.

Bug: 42387
Change-Id: I03e497b9bd4d693b9f7f7f674bc19372b532380f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436240
Reviewed-by: Filip Filmar <fmil@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Testability-Review: George Kulakowski <kulakowski@google.com>