[scenic] Scenic Stressor v0
This CL proposes a stressor for scenic using OpaqueTest.

We can start scenic + root_presenter in an isolated component
manager, provide a fake display and spam input commands onto a
basic view (simple_app).

In v0, the stressor sends 50 taps to a 640x480 touchscreen
at the center of the screen for 5 seconds. In response, the display
should render new rectangle colors on each tap.

A rudimentary mock display driver was written in Rust to
drive the fake display.

This test is configured to run on AEMU environments only
because Vulkan is required for Scenic to work.

Test: fx test scenic-stress-tests
Change-Id: I56de34ed225c01c1f1762b5a94594f4bc356d7b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429235
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Josh Gargus <jjosh@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>
Testability-Review: Fady Samuel <fsamuel@google.com>