[kernel][arm64][spinlock] minor tweak to the spinlock routine
In the inner spinlock loop, only loop back to the WFE if we detected
that the spinlock was already held by loading it and testing against
zero. If we fail to store because of an exclusive lock failure loop back
to the load instead of going through a WFE. Call this the two label
solution.

For all practical purposes the single label vs two label solution should
be the same but thinking about it the two label is a bit more logically
correct. Loop hard while you're failing a race with other cpus, WFE only
once you've lost.

Note there are conflicting examples that our spinlock is based on.

In the ARMv8 ARM version F.c in section K11.3.1 it describes a spinlock
implementation that matches ours with the exception of using 32bit
values instead of 64bit.

However, the ARM Trusted firmware at
https://github.com/ARM-software/arm-trusted-firmware/blob/master/lib/locks/exclusive/aarch64/spinlock.S#L50
at the time of writing uses the two label approach, which seems more
correct. Following this one as a golden example for now.

Bug: 56818
Change-Id: Ice24c3d8a375b592317d3e5f5d3ef9d8086b7840
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435150
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>