[chromedriver] Add outputs field to target.
This will allow it to get copied to a path within the out dir. The
attached bug was caused because the path to the chromedriver started
with "../.." because it was relative to the out dir and that was causing
issues with the upload path when we were trying to upload it with the
artifacts through artifactory.

Bug: 61695
Change-Id: I5c2d48552cc045a7debf85812b5842e5b646d348
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437072
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Fabrice de Gans-Riberi <fdegans@google.com>
Reviewed-by: Kevin Marshall <marshallk@google.com>