[ui-input] input-synthesis: move synthesizer tests to submodule
Move the existing synthesizer tests into a submodule, to prepare
for adding a set of new tests which exercise a different behavior
of the synthesizer.

This change doesn't affect the semantics of the existing tests.
It just helps us group things in the source code (as the number of
tests increase), and helps us more easily parse test output
(thanks to the grouping).

Bug: 60789
Test: fx test input-sythensis-tests
Change-Id: I9251488d0a80e04c53b2b16c8e32db090c279781
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434929
Commit-Queue: Mukesh Agrawal <quiche@google.com>
Testability-Review: Alice Neels <neelsa@google.com>
Reviewed-by: Viktar Donich <viktard@google.com>