[gidl][walker] Separate FIDL generation per GIDL file
Currently, the walker benchmarks are built within the same
source set, with each benchmark importing a large file
with all of the FIDL headers. This follows what is done
for the rust benchmarks so a minimal number of FIDL files
is used for each GIDL file.

This does not have significant effects on build time, but
it does move towards a leaner build configuration and the
output filenames now match the specific GIDL test -- making
it easier to understand what is going on.

This is the walker equivalent of a recent LLCPP change.

Change-Id: Ic98ffb83cdf17b835efaa8dd1c790a1dd5c703de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436628
Reviewed-by: Mitchell Kember <mkember@google.com>
Testability-Review: Mitchell Kember <mkember@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>