[garnet] Add fuchsia.media.Audio to SystemServices
Because fuchsia.media.Audio has been migrated to audio_core.cmx, it must
be allowed as a system-service for run-test-component environments,
such that it can be accessed through the system AudioCore.

Before, fuchsia.media.Audio was mapped as an injected-service through
audio.cmx. It can no longer be mapped as an injected-service, because we
do not want to inject another instance of audio_core.cmx along with the
existing AudioCore system-service, so we must allow fuchsia.media.Audio
to be a system-service.

Bug: 57744

Test: no behavior change
Change-Id: I69bb2b37a3f956eb1154d8cfd3560e6ccedb7cbd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436682
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Testability-Review: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: Payton Bradsky <pbradsky@google.com>