[SetUI][Storage] Refactor serialization.
SettingService structs are currently serialized by
device_storage and store_accessor. serde methods
are called directly on the types, which implement
Serialize and Deserialize.

This interaction limits the type's participation in
how it is serialized. This changelist moves this
behavior to be part of the DeviceStorageCompatible
trait with a default implementation matching the
current behavior. This will give flexibility in
the future for types to implement their own
serialization behavior.

Bug: 61061
Test: fx test -o setui_service_tests
Change-Id: I68435b62a2bfa853002d6f868a7ef1fe53673733
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436994
Reviewed-by: William Xiao <wxyz@google.com>
Testability-Review: William Xiao <wxyz@google.com>
Commit-Queue: Bryce Lee <brycelee@google.com>