[storage][extractor] Add extractor library 4/n
The library in its current state provides a way to write an image
file. Storage software can use this library to dump the data that is
relevant to them.

The library is in rust. It provides C bindings(in c-bindings directory)
that can be used in C++ code base.

This patch adds C bindings for image writer part of the library.

Test: Ran host side unit tests

Change-Id: I1259e53a48294d14be0d8d16eeed85d2bb1a4016
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432637
Commit-Queue: Vikram Auradkar <auradkar@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>