[build][alln] Move toolchain runtime deps to the Fuchsia build.
This ensures that the binaries.json entry for each toolchain runtime
library (e.g. libc++.so for C++ code, or libstd.so for Rust one)
is defined directly in the Fuchsia build, instead of relying on the
same entry being defined in out/default.zircon/binaries.json and
then later imported into the Fuchsia build version of that file.

Bug: 59916
Testing: Should not change the build.

Change-Id: I88fd7d5288666140a75cd334c5ac222004575f4c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436345
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>