[security][appmgr] Add system-updater allowlist
Add a new allowlist that will be used to restrict access to the
fuchsia.update.installer.Installer protocol to just those components
that should be able to initiate the installation of an update (and
tests).

A future change will use this allowlist in appmgr to enforce access.

Bug: 61800
Change-Id: I5f39623c4665b88a4cef20a7506c4b4990a62a77
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438177
Reviewed-by: Drew Fisher <zarvox@google.com>
Commit-Queue: Kevin Wells <kevinwells@google.com>