[netdevice] Negotiate rx signalling threshold
The signalling threshold to provide the device with more buffers used to
be hard-coded to half of the FIFO depth, which may not be ideal for
devices that operate at different speeds or have different burst traffic
capabilities. This change provides a mechanism for device implementers
to tune that option.

Fixed: 44835
Change-Id: I3890ef4270b5686794fd84f9421bcbe3ef8752f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/433858
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>
Testability-Review: Tamir Duberstein <tamird@google.com>