[scrutiny] Move rest service out of framework into app.
The Scrutiny command line is an interface to the framework and is part
of the application crate/binary. The rest network interface to Scrutiny
should also be in the app crate keeping the workings of the decoupled
agnostic from any particular interface.

Change-Id: I4e4ea691eea7553e84e4728cc4d6757a358ca90a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437200
Commit-Queue: Benjamin Wright <benwright@google.com>
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Testability-Review: Bryan Henry <bryanhenry@google.com>