[fvm] Remove some FVM header helper functions.
PartitionTableOffset
PartitionTableByteSizeForUsablePartitions
AllocTableOffset
AllocTableLengthForDiskSize
MetadataSizeForDiskSize
BackupStart
SlicesStart

Renamed AllocTableLengthForUsableSliceCount to
AllocTableByteSizeForUsableSliceCount() for clarity.

Some of the computations in fvm-host-test.cc are simplified
dramatically. This is a result of inlining the implementation of
ComputeRequiredDataSize and then simplifying the math expression. Most
of the terms fall out. The code seemed to think it was testing much more
than it was in practice so there should actually be no difference in
coverage: it was just validating the file size matched the FvmContainer
size.

Change-Id: I0b4e0a8454c391692cc6258876422d48a63bb305
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438434
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Testability-Review: James Sullivan <jfsulliv@google.com>