[ui-input] input synthesis: make register_device a function
We want `register_device!()` to have different behavior depending
on which injection protocol(s) are present on the system, so
this functionality will be provided by a function in an `impl`
of the `Injector` trait.

As the first step towards that, transform `register_device!()`
from a macro into a function.

Bug: 60789
Test: fx test input-synthesis-tests
Change-Id: I3922705a744c85b3d23cb873e343ec22c35052da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434931
Commit-Queue: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Testability-Review: Kevin Lindkvist <lindkvist@google.com>