[sysmem] Work around leak during fuzzer destructor
On device unbind the channel error handler is run, which causes all the
VMOs to be released. However, since the loop is being torn down
TrackedParentVmo.zero_children_wait_ is canceled, which means that
TrackedParentVmo continues to hold a reference to the
LogicalBufferCollection and that causes a memory leak.

Tearing down sysmem doesn't happen on a production system, so we can
just add a workaround for the fuzzer. If we run the loop until idle
after closing the VMO and channel handles then everything should work
properly.

Bug: 61392

Change-Id: I3a9db7c4ef87b5967e859efd6cb8cbf49be2ee5e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438268
Reviewed-by: John Rosasco <rosasco@google.com>
Testability-Review: John Rosasco <rosasco@google.com>
Commit-Queue: John Bauman <jbauman@google.com>