[fidl] Temporarily disable fidlgen_hlcpp_test tests
clang-format is updated as a part of the clang roll and produces a
different output from what is expected in the golden file for
TestCodegenSource/union.test.json. To facilitate the roll, we will
temporarily disable this test, then turn it back on and regenerate the
tests.

Bug: 61989
Change-Id: I48bf37912c5f9e17b71b957207d8c6e7282b39d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438974
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>